@@ -14,11 +14,13 @@
   "dependencies": {
     "@emailjs/browser": "^3.11.0",
     "@hookform/resolvers": "^3.3.2",
+    "@stripe/react-stripe-js": "^2.3.1",
+    "@stripe/stripe-js": "^2.1.10",
     "github-slugger": "^2.0.0",
     "gray-matter": "^4.0.3",
     "marked": "^4.2.4",
     "micro": "^10.0.1",
-    "next": "13.0.6",
+    "next": "^13.5.6",
     "next-mdx-remote": "^4.2.0",
     "preline": "^1.9.0",
     "react": "18.2.0",
