@@ -1,15 +1,40 @@
 import { useState, useEffect } from "react";
 import ReservationCheckoutFlowStep1 from "@layouts/partials/ReservationCheckoutFlowStep1";
 import ReservationCheckoutFlowStep2 from "@layouts/partials/ReservationCheckoutFlowStep2";
+import ReservationCheckoutFlowStep3 from "@layouts/partials/ReservationCheckoutFlowStep3";
+import emailjs from '@emailjs/browser';
+import { emailConfig } from '@config/emailConfig';
 
 const Form = ({ closeReservationCheckout }) => {
 
+  const options = {
+    // passing the client secret obtained from the server
+    clientSecret: "{{CLIENT_SECRET}}",
+  };
   const [step, setStep] = useState(1);
   const [formData, setFormData] = useState({});
+  const { emailServiceId, emailTemplateId, emailPublicKey } = emailConfig;
 
   useEffect(() => {
-    console.log(formData); // This will log the updated formData
-  }, [formData]);
+
+    if (step === 3) {
+      const templateParams = {
+        from_name: formData.firstName + ' ' + formData.lastName,
+        from_email: formData.email,
+        from_phone: formData.phone,
+        message: `${formData.firstName} ${formData.lastName} has joined the waitlist! They are ${formData.isPregnant ? 'currently pregnant' : 'not currently pregnant'}. Their due date is ${formData.dueDate}. They would like to visit from ${formData.desiredVisitDates[0]} to ${formData.desiredVisitDates[1]}. They ${formData.joinMailingList ? 'would' : 'would not'} like to join the mailing list.`,
+      };
+
+      emailjs.send(emailServiceId, emailTemplateId, templateParams, emailPublicKey)
+        .then((result) => {
+          console.log(result.text);
+        })
+        .catch((error) => {
+          console.log(error.text);
+        });
+    }
+  });
+
 
   const handleNextStep = (data) => {
     setFormData({ ...formData, ...data });
@@ -26,7 +51,7 @@ const Form = ({ closeReservationCheckout }) => {
   };
 
   return (
-    <section className="shadow">
+    <section className="p-2">
       {step === 1 && (
         <ReservationCheckoutFlowStep1
           onSubmit={handleNextStep}
@@ -43,7 +68,7 @@ const Form = ({ closeReservationCheckout }) => {
         />
       )}
       {step === 3 && (
-        <ReservationCheckoutFlowStep2
+        <ReservationCheckoutFlowStep3
           onSubmit={handleNextStep}
           onPrev={handlePrevStep}
           onClose={onClose}
